Rework utilities
