+++ operator/internal/autoscaler/webhook.go
@@ -1,4 +1,4 @@
-package utils
+package autoscaler
 
 import (
 	"bytes"
