+++ operator/internal/builders/pod.go
@@ -1,4 +1,4 @@
-package utils
+package builders
 
 import (
 	"os"
