+++ operator/internal/builders/suite_test.go
@@ -1,0 +1,22 @@
+package builders
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+)
+
+func TestBuilders(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Builders Suite")
+}
+
+var _ = BeforeSuite(func() {
+	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
+})
+
+var _ = AfterSuite(func() {})
