+++ operator/internal/utils/gametype.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/MirrorStudios/fallernetes-operator/api/v1alpha1"
 	"github.com/go-logr/logr"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -44,26 +43,4 @@ func GetNewestFleet(fleets []v1alpha1.Fleet) *v1alpha1.Fleet {
 		}
 	}
 	return newest
-}
-
-func GetFleetObjectForType(gametype *v1alpha1.GameType) *v1alpha1.Fleet {
-	labels := gametype.Labels
-	if labels == nil {
-		labels = map[string]string{}
-	}
-	labels["gametype"] = gametype.Name
-
-	fleet := &v1alpha1.Fleet{
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName: gametype.Name + "-",
-			Namespace:    gametype.Namespace,
-			Labels:       labels,
-			OwnerReferences: []metav1.OwnerReference{
-				*metav1.NewControllerRef(gametype, v1alpha1.GroupVersion.WithKind("GameType")),
-			},
-		},
-		Spec: gametype.Spec.FleetSpec,
-	}
-
-	return fleet
 }
